commerce_intel_project/requirements.txt
streamlit
pandas
plotly
python-dotenv
openpyxl
groq\n